requirements.txt
python-dotenv
slack_bolt
anthropic
certifi
aiohttp

# Development dependencies
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pre-commit==3.5.0
autopep8==2.0.4